requirements.txt
streamlit
pymupdf
langchain
langchain-community
faiss-cpu
google-generativeai
sentence-transformers
